backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt --strip-extras requirements.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
blessed==1.20.0
    # via -r requirements.in
boto3==1.34.72
    # via
    #   -r requirements.in
    #   django-ses
botocore==1.34.72
    # via
    #   boto3
    #   s3transfer
croniter==2.0.3
    # via -r requirements.in
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20230430
    # via django-import-export
dj-database-url==2.1.0
    # via -r requirements.in
django==4.2.11
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-cors-headers
    #   django-import-export
    #   django-jazzmin
    #   django-jquery
    #   django-picklefield
    #   django-q2
    #   django-ses
    #   django-storages
    #   djangorestframework
    #   drf-spectacular
django-cors-headers==4.3.1
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
django-import-export==3.3.7
    # via -r requirements.in
django-jazzmin==2.6.1
    # via -r requirements.in
django-jquery==3.1.0
    # via -r requirements.in
django-picklefield==3.1
    # via django-q2
django-q2==1.6.2
    # via -r requirements.in
django-ses==3.5.2
    # via -r requirements.in
django-storages==1.14.2
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   drf-spectacular
drf-spectacular==0.27.1
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
gevent==24.2.1
    # via -r requirements.in
greenlet==3.0.3
    # via gevent
gunicorn==21.2.0
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.21.1
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
markuppy==1.14
    # via tablib
odfpy==1.4.1
    # via tablib
openpyxl==3.1.2
    # via tablib
packaging==24.0
    # via gunicorn
pillow==10.2.0
    # via -r requirements.in
psutil==5.9.8
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   croniter
pytz==2024.1
    # via
    #   croniter
    #   django-ses
    #   djangorestframework
pyyaml==6.0.1
    # via
    #   drf-spectacular
    #   tablib
referencing==0.34.0
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
s3transfer==0.10.1
    # via boto3
six==1.16.0
    # via
    #   blessed
    #   python-dateutil
sqlparse==0.4.4
    # via django
tablib==3.5.0
    # via
    #   -r requirements.in
    #   django-import-export
typing-extensions==4.10.0
    # via dj-database-url
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.2.1
    # via botocore
wcwidth==0.2.13
    # via blessed
whitenoise==6.6.0
    # via -r requirements.in
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib
zope-event==5.0
    # via gevent
zope-interface==6.2
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
